backend/requirements.txt
fastapi
uvicorn
pydub
elevenlabs
pydantic
google-cloud
google-generativeai
MLB-StatsAPI
requests
mutagen
redis
google-cloud-storage
google-cloud-firestore
google-cloud-aiplatform